requirements.txt
alembic==1.4.2
attrs==19.3.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
idna==2.10
cryptography>=36.0.1
cffi>=1.12
pycparser==2.21
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
PyJWT==1.7.1
PyMySQL==0.10.0
pyparsing==2.4.7
pytest==6.0.1
pytest-asyncio==0.14.0
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
urllib3==1.25.10
uWSGI==2.0.19.1
Werkzeug==1.0.1
zipp==3.1.0
